requirements.txt
pip~=24.0
Django~=5.1.3
asgiref~=3.8.1
sqlparse~=0.5.2
setuptools~=65.5.0
pillow~=11.0.0
pyowm~=3.3.0
geojson~=2.5.0
requests~=2.32.3
psycopg2~=2.9.10
certifi~=2024.12.14
urllib3~=2.2.3
idna~=3.10
djangorestframework~=3.15.2
django-leaflet~=0.31.0
boto3~=1.36.6
botocore~=1.36.6
yandexcloud~=0.331.0
environ~=1.0
dj-config-url~=0.1.1
django-storages